+++ base/array.jl
@@ -2505,6 +2505,8 @@ Return `nothing` if there is no such element.
 Indices or keys are of the same type as those returned by [`keys(A)`](@ref)
 and [`pairs(A)`](@ref).
 
+See also [`filterfirst`](@ref) to get the element instead of its index.
+
 # Examples
 ```jldoctest
 julia> A = [1, 4, 2, 2]
@@ -3089,6 +3091,37 @@ function filter!(f, a::AbstractVector)
     end
     return a
 end
+
+"""
+    filterfirst(f, itr)
+
+Get the first element of an iterable collection `itr` for which the
+predicate `f` returns `true`. Throw an `ArgumentError` if there is no such
+element, like [`first`](@ref) does for an empty collection; returning
+`nothing` would be ambiguous, since the element itself may be `nothing`.
+
+See also [`findfirst`](@ref) for the index of that element,
+[`Iterators.filter`](@ref).
+
+!!! compat "Julia 1.14"
+    This method requires at least Julia 1.14.
+
+# Examples
+```jldoctest
+julia> filterfirst(>=(5), 2:2:10)
+6
+
+julia> filterfirst(iseven, [1; 2; 3; 4])
+2
+
+julia> filterfirst(isuppercase, "hello World")
+'W': ASCII/Unicode U+0057 (category Lu: Letter, uppercase)
+
+julia> filterfirst(isodd, 2:2:10)
+ERROR: ArgumentError: collection must be non-empty
+```
+"""
+filterfirst(f, itr) = first(Iterators.filter(f, itr))
 
 """
     filter(f)